% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmhash64.R
\name{FarmHash32Hex}
\alias{FarmHash32Hex}
\title{Computes the 32-bit FarmHash hash value of each string in the input vector
and returns the hexadecimal representation of the hash values.}
\usage{
FarmHash32Hex(strv)
}
\arguments{
\item{strv}{The input character vector containing the strings to be hashed.}
}
\description{
Computes the 32-bit FarmHash hash value of each string in the input vector
and returns the hexadecimal representation of the hash values.
}
